+++ src/components/date_picker/date_picker_range.test.tsx
@@ -14,11 +14,23 @@ import { EuiDatePickerRange } from './date_picker_range';
 import { EuiDatePicker } from './date_picker';
 
 describe('EuiDatePickerRange', () => {
-  test('is rendered', () => {
+  it('is rendered', () => {
     const component = render(
       <EuiDatePickerRange
         startDateControl={<EuiDatePicker />}
         endDateControl={<EuiDatePicker />}
+        {...requiredProps}
+      />
+    );
+
+    expect(component).toMatchSnapshot();
+  });
+
+  it('uses individual EuiDatePicker props', () => {
+    const component = render(
+      <EuiDatePickerRange
+        startDateControl={<EuiDatePicker className="hello" />}
+        endDateControl={<EuiDatePicker className="world" />}
         {...requiredProps}
       />
     );